pyproject.toml
[tool.poetry]
name = "mc-benchmark"
version = "0.1.0"
description = ""
authors = ["James McNeill <jpmmcneill@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.18"
numpy = "^1.26.3"
duckdb = "^0.9.2"
memory-profiler = "^0.61.0"
pandas = "^2.1.4"
plotly = "^5.18.0"
numba = "^0.58.1"
pyarrow = "^14.0.2"
pyyaml = "^6.0.1"
click = "^8.1.7"
jupyter = "^1.0.0"
dash = "^2.14.2"
polars = "^0.20.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
